C# Bindings documentation and location change
Added documentation to the C# Bindings for 2.0.0 Alpha.
AlertLayoutChangeWrapper is undocumented, there are not enough public api yet to differentiate it from normal Alert.

Location change:
Moved from bindings/... to extensions/...
In this case extensions/csharp

Include csharp directory in harmoniaui directory or next to .gdextension file

For ease of development in C# I'm developing harmoniaui-editor

Take care!
2.0.0 Alpha C# bindings.
